Typecheck and other comment fixes
