+++ src/views/AnalysisModule/Analysis/styles.scss
@@ -6,31 +6,6 @@
     border-radius: var(--dui-border-radius-card);
     background-color: var(--color-foreground);
 
-    .content {
-        .content-item {
-            margin: var(--spacing-medium) 0;
-
-            .sub-heading {
-                color: var(--color-text-placeholder);
-            }
-
-            .pillar-list-content {
-                display: flex;
-
-                .analyst {
-                    flex-basis: 1;
-                    margin-right: var(--spacing-small);
-                    font-weight: var(--font-weight-bold);
-                }
-
-                .analysis-pillar {
-                    flex-basis: 1;
-                    margin-left: var(--spacing-small);
-                }
-            }
-        }
-    }
-
     .header-right {
         display: flex;
         align-items: center;
@@ -40,91 +15,70 @@
             margin: 0 var(--dui-spacing-medium);
         }
     }
-    .bold-text {
-        font-weight: var(--font-weight-bold);
-    }
 
-    .body {
-        display: flex;
-        align-items: flex-start;
-        padding: var(--dui-spacing-small);
-    }
     .pillar-content {
         padding: 0 !important;
 
         .content {
             padding: var(--dui-spacing-large);
-        }
-    }
 
-    .pillar-analyses {
-        display: flex;
-        flex-direction: column;
-        border-left: 4px solid var(--color-separator);
-
-        .accordion-button {
-            display: flex;
-            flex-grow: 1;
-            justify-content: flex-start;
-            outline: var(--color-separator);
-            border: var(--width-separator-thin) solid var(--color-separator);
-            border-left: none;
-            border-radius: 0;
-            background: none;
-            padding: var(--dui-spacing-extra-large) var(--dui-spacing-large);
-            color: var(--color-link);
-        }
-
-        .pillar-analysis-content {
-            padding: var(--dui-spacing-large);
-        }
-        .pillar {
-            display: flex;
-            flex-grow: 1;
-            margin: var(--spacing-small) 0;
-            background: white;
-
-            &:not(:last-child) {
-                border-bottom: var(--width-separator-thin) solid var(--color-separator);
-            }
-
-            .heading {
-                display: flex;
-
-                .left {
-                    display: flex;
-
-                    .title {
-                        display: flex;
-                        flex-grow: 1;
-                    }
-
-                    .tag {
-                        flex-shrink: 0;
-                        margin-left: var(--spacing-large);
-                    }
-                }
-            }
-
-            .sub-heading {
-                display: flex;
-            }
-
-            .item {
+            .content-item {
                 margin: var(--spacing-medium) 0;
 
-                .bold-text {
-                    display: block;
+                .sub-heading {
+                    display: flex;
+                    color: var(--color-text-placeholder);
+                }
+
+                .pillar-list-content {
+                    display: flex;
+
+                    .analyst {
+                        flex-basis: 1;
+                        margin-right: var(--spacing-small);
+                    }
+
+                    .analysis-pillar {
+                        flex-basis: 1;
+                        margin-left: var(--spacing-small);
+                    }
+                }
+            }
+        }
+        .pillar-analyses {
+            display: flex;
+            flex-direction: column;
+            border-left: 4px solid var(--color-separator);
+
+            .accordion-button {
+                display: flex;
+                flex-grow: 1;
+                justify-content: flex-start;
+                outline: var(--color-separator);
+                border: var(--width-separator-thin) solid var(--color-separator);
+                border-left: none;
+                border-radius: 0;
+                background: none;
+                padding: var(--dui-spacing-extra-large) var(--dui-spacing-large);
+                color: var(--color-link);
+            }
+
+            .pillar-analysis-content {
+                padding: var(--dui-spacing-large);
+
+                .pillar-list {
+                    display: flex;
+                    flex-grow: 1;
+
+                    .bold-text {
+                        margin-right: var(--spacing-small);
+                    }
                 }
             }
         }
     }
-    .pillar-list {
-        display: flex;
-        flex-grow: 1;
 
-        .bold-text {
-            margin-right: var(--spacing-small);
-        }
+    .bold-text {
+        font-weight: var(--font-weight-bold);
     }
 }
